New directions component
Created a new component and added all html and ts code that was previously implemented directly onto the page, onto the new component.
