+++ src/app/components/directions/directions.component.html
@@ -1,0 +1,28 @@
+<form id="form" (ngSubmit)="getDirections()">
+  <ion-toolbar class="toolbar" style="margin-top: 60px;">
+      <ion-item>
+          <ion-input [(ngModel)]="directions.start" (ionChange) = "displayShuttle()" name='start' placeholder="Start"></ion-input>
+      </ion-item>
+      <ion-item>
+          <ion-input [(ngModel)]="directions.destination" (ionChange) = "displayShuttle()" name='destination' placeholder="Destination"></ion-input>
+      </ion-item>
+      <ion-segment id="transportMode">
+          <ion-segment-button class="transportButtons" id="walking">
+            <ion-label>WALKING</ion-label>
+          </ion-segment-button>
+          <ion-segment-button class="transportButtons" id="driving">
+            <ion-label>DRIVING</ion-label>
+          </ion-segment-button>
+          <ion-segment-button class="transportButtons" id="transit">
+              <ion-label>PUBLIC TRANSIT</ion-label>
+          </ion-segment-button>
+          <ion-segment-button [style.display]="'none'" class="transportButtons" id="shuttle">
+            <ion-label>SHUTTLE BUS</ion-label>
+        </ion-segment-button>
+        </ion-segment>
+      <ion-button  type="submit" expand="full" (click)="getTransportation()">Get Directions</ion-button>
+  </ion-toolbar> 
+</form>
+
+<!-- Div to display step by step direction in text for any given route -->
+ <!-- <div id="directionsPanel"></div> -->
