+++ src/app/components/directions/directions.component.scss
@@ -1,0 +1,16 @@
+.transportButtons{
+    height: 50px;
+ }
+
+ #form{
+    
+ }
+
+ #directionsPanel{
+    overflow: auto;
+    height: 400px;
+ }
+
+ #transportMode{
+    height: min-content;
+ }
