+++ src/app/components/new-route/new-route.component.scss
@@ -3,19 +3,3 @@
     width: 100%;
  }
  
- .transportButtons{
-    height: 50px;
- }
-
- #form{
-    
- }
-
- #directionsPanel{
-    overflow: auto;
-    height: 400px;
- }
-
- #transportMode{
-    height: min-content;
- }
